Add files via upload
In the process of curating the dataset for sentiment analysis of Bengali tweets, extensive data cleansing protocols were meticulously executed. This entailed the excision of superfluous URLs, the eradication of emotive pictograms, and the systematic purification of textual anomalies. By rigorously refining the data corpus, we ensured the retention of semantically relevant content, thereby augmenting the reliability and accuracy of subsequent sentiment analysis endeavors.